Simplify debug screens layout and update RGB model
- Refactor `InfDebugScreen`, `CpsDebugScreen`, `VcnDebugScreen`, `PalDebugScreen`, `MazDebugScreen`, and `VmpDebugScreen` to remove responsive landscape/portrait logic.
- Replace `BoxWithConstraints` with a standard `Column` layout in debug screens, stacking the list and detail views vertically.
- Update `RGB` data model to use `Int` instead of `UByte` for color components.
- Update `PalRepository` to return `Int` from `convert6bitTo8bit` to match the model change.

+++ composeApp/src/commonMain/kotlin/pl/pelotasplus/eyeofbeholder/data/repository/PalRepository.kt
@@ -35,8 +35,8 @@ class PalRepositoryImpl(
         return resourceRepository.listResources(".PAL")
     }
 
-    private fun convert6bitTo8bit(byte: UByte): UByte {
-        return ((byte.toInt() * 255) / 63).toUByte()
+    private fun convert6bitTo8bit(byte: UByte): Int {
+        return ((byte.toInt() * 255) / 63)
     }
 
     companion object {
